♻️ Refactor CommandStep utility functions
