+++ modules/fantasy-battle/src/Commands/bio.ts
@@ -3,12 +3,12 @@ import { Command, RegexCommand } from "@discord-bot/create-client"
 import { ResourceDocument } from "../Models/PcResource"
 import { Attribute } from "../Models/PlayerCharacter/types"
 
-import parseFlags, { FlagsObject } from "../Utils/parseArgs"
-import rejectIfNotPlayerOrDm from "../Utils/rejectIfNotPlayerOrDm"
+import parseFlags, { FlagsObject } from "../Utils/CommandStep/parseArgs"
+import rejectIfNotPlayerOrDm from "../Utils/CommandStep/rejectIfNotPlayerOrDm"
 import { commandWithFlags } from "../Utils/regex"
-import { getPlayerUser } from "../Utils/getUser"
+import { getPlayerUser } from "../Utils/CommandStep/getUser"
 import { logSuccess } from "../Utils/commandLog"
-import getPlayerChar from "../Utils/getPlayerChar"
+import getPlayerChar from "../Utils/CommandStep/getPlayerChar"
 
 export const test: RegexCommand.test = commandWithFlags(
   /bio/,
