+++ modules/fantasy-battle/src/Commands/rollDmg.ts
@@ -4,12 +4,12 @@ import { capture, concat, fromList, optional, optionalSpace } from "@discord-bot
 
 import { attributeNameRegex, getAttributeByNickname } from "../Models/PlayerCharacter/helpers/attributes"
 
-import parseFlags, { FlagsObject } from "../Utils/parseArgs"
-import rejectIfNotPlayerOrDm from "../Utils/rejectIfNotPlayerOrDm"
+import parseFlags, { FlagsObject } from "../Utils/CommandStep/parseArgs"
+import rejectIfNotPlayerOrDm from "../Utils/CommandStep/rejectIfNotPlayerOrDm"
 import { commandWithFlags } from "../Utils/regex"
-import { getPlayerUser } from "../Utils/getUser"
+import { getPlayerUser } from "../Utils/CommandStep/getUser"
 import { logSuccess } from "../Utils/commandLog"
-import getPlayerChar from "../Utils/getPlayerChar"
+import getPlayerChar from "../Utils/CommandStep/getPlayerChar"
 
 export const test: RegexCommand.test = commandWithFlags(
   concat(
